LoggerTest: inlined to NullLoggerTest
unfortunately, due to limitations of C++, it will not be reusable after all…

When I pass an object inheriting after Logger (for example NullLogger) -> the methods from Logger are called, not those from NullLogger!

The solution would be to use more macros (ugly). Plus the error reporting would be ugly. So, we end up with a bit of copy–paste. Great language! :P