Storing string literals in FLASH (PROGMEM) instead of RAM.
PROGMEM macro can be used to mark global/static vars as to be stored in FLASH.

PSTR() macro allows to store string literals in FLASH, within code blocks.
It actually creates a helper global/static var with PROGMEM attribute.

F() additionally casts it to `__FlashStringHelper`,
so that Arduino internally is able to recognize,
that it is a FLASH-stored value, to be able to interact with it with dedicated `_P` functions.

OPEN QUESTION:
  `__FlashStringHelper` is there so that Arduino can recognize that the value is in FLASH
  and use special functions to interact with it.
  Why is that my format() function works with PSTR() then?!
  It does not use any special functions to access FLASH data…
  But it works anyway and RAM usage is reduced :]
  * https://forum.arduino.cc/t/esp8266-f-macro-is-unnecessary-on-esp-chips/1075805/1
  * https://community.platformio.org/t/esp8266-arduino-f-macro-is-unnecessary-on-esp-chips/31399
  * https://discord.com/channels/583251190591258624/1062392340783829063/1062392340783829063
  After all, it looks like `F()` is unnecessary on ESP8266.

advantage: the string literals are NOT loaded into RAM at "startup" time – they are read directly from FLASH
disadvantage (?): possibly a bit slower (?) – but I don't know how could I notice that -> so no problem
disadvantage: need to remember to put all the string literals into the macro(s) -> well, this sucks, but what can I do ;p

Convention:
* `PROGMEM` for global/static vars
* `PSTR` when `const char *` is accepted
* `F` when `String` or `__FlashStringHelper` is accepted

https://github.com/esp8266/Arduino/blob/master/doc/PROGMEM.rst

Before:
RAM:   [======    ]  57.2% (used 46824 bytes from 81920 bytes)
Flash: [======    ]  62.2% (used 649980 bytes from 1044464 bytes)

After:
RAM:   [=====     ]  52.8% (used 43220 bytes from 81920 bytes)
Flash: [======    ]  62.3% (used 650368 bytes from 1044464 bytes)

Difference?
46824-43220 = ~3,6 KB less RAM
100-43220/46824*100 = ~7,7% less RAM
57,2-52.8 = ~4,4 percent points less RAM

Not really a big difference.

But ofc, the more and the longer the strings, the more is to gain.
In some edge cases, this could make the difference.
And on a device with less RAM, this would make all the difference.
On ESP8266, it is nice, but maybe not really necessary.