removed/changed todo: (not) using array instead of vector (yet)
It's not possible to declare that I want an array WITHOUT declaring its size at the same time.
It can be worked–around by abstracting the type into a template:

```cpp
template<int LOGGERS_COUNT>
class SequenceLogger final: Logger {
    using Loggers = std::array<std::reference_wrapper<Logger>, LOGGERS_COUNT>;
```

```cpp
const auto innerLoggers = std::to_array<std::reference_wrapper<Logger>>({
    stringLogger1,
    stringLogger2
});

SequenceLogger<2> logger{innerLoggers};
```

As can be seen, it kinda works: my type requires an array without specifying the size. But, the callers have to specify the size, making it harder/error–prone in the end.

SO, the question is: why would I want an array anyway? Actually, I only want to loop – I don't care if the size of the looped list is fixed or not.
So, actually, it would not be an improvement from the SequenceLogger design POV <<< the change from Vector to Array <<< it would still be "arbitrary" distinction over a thing we don't need to care about here.

SPAN is probably the answer :)